requirements.txt
openai-whisper
dashscope
gradio
pydub
pyannote.audio
# torch 不在这里装！请手动安装GPU版：
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
